requirements.txt
git+https://github.com/stephenmcd/mezzanine.git#egg=Mezzanine
django-extensions
django-phonenumber-field
django-logentry-admin
psycopg2
django-autoslug
pandas
ipython
geocoder==1.33.0
# To be replaced by bokeh graphs
plotly
numpy
bokeh==0.12.13
flexx
tornado
